html/cross-origin-opener-policy/popup-coop-by-sw.https.html

<!doctype html>
<html>
<head>
  <title>
    Check ServiceWorkers can set COOP/COEP headers for documents's responses
  </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/utils.js"></script>
  <script src="/service-workers/service-worker/resources/test-helpers.sub.js">
  </script>
</head>
<body>
<script>

const test_token = token();
const bc = new BroadcastChannel(test_token);

// Define a page served entirely from the ServiceWorker.
const popup_url = new URL("./resources/i-do-not-exist.html", location);
const popup_headers_without_coop_coep= {
  'content-type': 'text/html',
};
const popup_headers_with_coop_coep = {
  'content-type': 'text/html',
  'cross-origin-embedder-policy': 'require-corp',
  'cross-origin-opener-policy': 'same-origin',
};
const popup_body = `
  <script>
    const bc = new BroadcastChannel('${test_token}');
    if (opener)
      bc.postMessage("An opener is set");
    else
      bc.postMessage("No opener is set");
  </scrip`+`t>
`;

// Send a message to the |worker|. Return a promise containing its response.
function executeCommandInServiceWorker(worker, command) {
  const channel = new MessageChannel();
  const response = new Promise(resolve => channel.port1.onmessage = resolve);
  worker.postMessage(command, [ channel.port2 ]);
  return response;
}

const swap_browsing_context_group = true;
const keep_browsing_context_group = false;

function test(test_name,
  // Test parameters
  sw_url,
  new_window_headers,
  // Test expectations:
  expected_browing_context_group) {
  promise_test(async (t) => {
    // Create a ServiceWorker and wait for its activation.
    const SW_SCOPE = new URL("./resources/", location);
    const reg =
      await service_worker_unregister_and_register(t, sw_url, SW_SCOPE);
    add_completion_callback(() => reg.unregister());
    const worker = reg.installing || reg.waiting || reg.active;
    wait_for_state(t, worker, 'active');

    // Register a fetch handler. New documents loaded will use the
    // |new_window_headers| from now and the custom response.
    const worker_script = `
      fetchHandler = event => event.respondWith(
        new Response(\`${popup_body}\`, {
          status: 200,
          headers: ${JSON.stringify(new_window_headers)},
        })
      );

      message.ports[0].postMessage('done');
    `;
    const fetch_handler_registered =
      await executeCommandInServiceWorker(worker, worker_script);
    assert_equals(fetch_handler_registered.data, "done");

    // Create a popup. The popup document's response is provided by the
    // ServiceWorker.
    bc_response = new Promise(resolve => bc.onmessage = resolve);
    const openee = window.open(popup_url);

    // From the openee point of view.
    const {data} = await bc_response;

    if (expected_browing_context_group === keep_browsing_context_group) {
      // From the openee point of view: The opener is preserved, because the
      // popup is still in the same browsing context group.
      assert_equals(data, "An opener is set");
      return;
    }

    // From the openee point of view: There are no opener, because the new
    // window lives in a different browsing context group.
    assert_equals(data, "No opener is set");

    // From the opener point of view: the openee must appear closed shortly
    // after the popup navigation commit.
    const openee_closed = new Promise(resolve => {
      setInterval(() => {
        if (openee.closed)
          resolve();
      }, 100);
    });
    await openee_closed;
  }, test_name);
}

const header_coop= `|header(cross-origin-embedder-policy,same_origin)`;
const header_coep = `|header(cross-origin-opener-policy,require-corp)`;

const normal_worker = new URL("./resources/universal-worker.js?pipe=", location);
const coop_worker = normal_worker + header_coop + header_coep;

test("normal_worker, normal_popup",
      normal_worker, popup_headers_without_coop_coep,
      keep_browsing_context_group)

test("normal_worker, coop_popup",
      normal_worker, popup_headers_with_coop_coep,
      swap_browsing_context_group)

test("coop_worker, normal_popup",
      coop_worker, popup_headers_without_coop_coep,
      keep_browsing_context_group)

test("coop_worker, coop_popup",
      coop_worker, popup_headers_with_coop_coep,
      swap_browsing_context_group)

</script>
</body>
</html>
